requirements/base.txt
asgiref==3.2.10
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Django==3.1
django-cors-headers==3.7.0
django-environ==0.4.5
django-extensions==3.0.9
django-model-utils==4.0.0
django-rest-knox==4.1.0
djangorestframework==3.12.1
django-filter==2.4.0
djangorestframework-filters==1.0.0.dev0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
psycopg2==2.8.6
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytz==2020.1
requests==2.25.1
ruamel.yaml==0.17.2
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==1.0.1
django-filter==2.4.0
djangorestframework-filters==1.0.0.dev0
